+++ app/Services/PdfMaker/PdfMaker.php
@@ -1,6 +1,16 @@
 <?php
 
-namespace Tests\Feature\PdfMaker;
+/**
+ * Invoice Ninja (https://invoiceninja.com)
+ *
+ * @link https://github.com/invoiceninja/invoiceninja source repository
+ *
+ * @copyright Copyright (c) 2020. Invoice Ninja LLC (https://invoiceninja.com)
+ *
+ * @license https://opensource.org/licenses/AAL
+ */
+
+namespace App\Services\PdfMaker;
 
 class PdfMaker
 {
